requirements.txt
pyslowfast
decord
lightning==2.2.5
scikit-learn==1.5.1
torch==2.3.0
torchvision==0.18.0
